Taro moves N squares numbered from 1 to N.
Next, we use a variable current to keep track of the current square Taro is on. We then loop k times to make k moves. For each move, we first check the next square to move to using to[current-1][0]. We subtract 1 from current because the squares are numbered from 1 to n, but our to array is 0-indexed.

If the next square is -1, we move back to the previous square by subtracting 1 from current. Otherwise, we move to the next square by setting current to next_square.

After each move, we need to update the to array to remove the current square from each row. We do this by calling row.shift on each row in to.

Finally, we output the current square using puts current.

This code should work for any input that satisfies the given conditions.